refactor: reorg boot of host & iframe
