+++ core/elements/src/isolation/boot/host/assets.ts
