+++ core/elements/src/isolation/boot/index.ts
@@ -1,8 +1,14 @@
 /**
- * Host-side setup: how an isolation iframe gets born. `bootIsolation` runs the
- * boot sequence and returns a handle; the leaf modules (assets, styles,
- * import-map, spec) are the self-contained reads it composes.
+ * How an isolation iframe gets born. `bootIsolation` runs the boot sequence and
+ * returns a handle.
+ *
+ * The subdirectories mark the **trust boundary** by realm of execution:
+ *  - `host/` — code that runs in the trusted host. `boot.ts` is the director;
+ *    `assets`/`styles`/`import-map`/`spec` are the reads it composes; `srcdoc.ts`
+ *    assembles the iframe's HTML (it serializes the iframe code but runs host-side).
+ *  - `iframe/` — code that runs INSIDE the untrusted sandbox (`main.ts`). It never
+ *    executes in the host; `host/srcdoc.ts` embeds it via `.toString()`.
  */
 
-export { bootIsolation, type IsolationHandle } from "./boot.js";
-export { specsEqual } from "./spec.js";
+export { bootIsolation, type IsolationHandle } from "./host/boot.js";
+export { specsEqual } from "./host/spec.js";
